# backend/src/main/resources/swagger.yml
openapi: 3.0.3
info:
  title: Aluguel de Carros API
  description: |
    API REST do serviço de aluguel de carros (Micronaut).

    **Autenticação:** a maioria das rotas exige header `Authorization: Bearer <JWT>`.

    **Rotas públicas (sem token):**
    - `POST /auth/login`
    - `POST /clientes`
    - `POST /bancos`
    - `POST /empresas`
    - `GET /swagger-ui`
    - `GET /swagger-ui/swagger.yml`
  version: 0.1.0
servers:
  - url: http://localhost:8080/api/aluguelcarros/v1
    description: Ambiente local
...
paths:
...
  /pedidos:
    post:
      tags: [Pedidos]
      summary: Cadastrar pedido
      description: |
        Ao criar um pedido:
        - `status` é definido como `EM_ANALISE`
        - `dataFim` deve ser maior que `dataInicio`
        - o automóvel deve estar com status `DISPONIVEL`
      operationId: cadastrarPedido
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoRequest'
            example:
              clienteId: 1
              automovelId: 1
              dataInicio: "2026-04-15"
              dataFim: "2026-04-20"
              status: "APROVADO"
      responses:
        '200':
          description: Pedido criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidoResponse'
        '400':
          description: Erro de validação
        '401':
          description: Não autorizado
        '422':
          description: Regra de negócio violada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensagemErro'
    get:
      tags: [Pedidos]
      summary: Listar pedidos
      operationId: listarPedidos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PedidoResponse'
        '401':
          description: Não autorizado
  /pedidos/{id}:
    parameters:
      - $ref: '#/components/parameters/PedidoId'
    get:
      tags: [Pedidos]
      summary: Buscar pedido por id
      operationId: buscarPedidoPorId
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidoResponse'
        '401':
          description: Não autorizado
        '404':
          description: Pedido não encontrado
    put:
      tags: [Pedidos]
      summary: Atualizar pedido
      operationId: atualizarPedido
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoRequest'
      responses:
        '200':
          description: Pedido atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidoResponse'
        '400':
          description: Erro de validação
        '401':
          description: Não autorizado
        '404':
          description: Pedido não encontrado
        '422':
          description: Regra de negócio violada
    delete:
      tags: [Pedidos]
      summary: Remover pedido
      operationId: removerPedido
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Pedido removido
        '401':
          description: Não autorizado
        '404':
          description: Pedido não encontrado
  /pedidos/status/{status}:
    parameters:
      - $ref: '#/components/parameters/PedidoStatusPath'
    get:
      tags: [Pedidos]
      summary: Listar pedidos por status
      operationId: listarPedidosPorStatus
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de pedidos filtrada por status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PedidoResponse'
        '401':
          description: Não autorizado
        '422':
          description: Status inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensagemErro'
  /pedidos/me:
    get:
      tags: [Fluxos de Negócio]
      summary: Listar pedidos do cliente autenticado
      description: Disponível apenas para perfil `cliente`.
      operationId: listarPedidosMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de pedidos do cliente autenticado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PedidoResponse'
        '401':
          description: Não autorizado
        '422':
          description: Perfil sem permissão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensagemErro'
  /pedidos/{id}/cancelar:
    parameters:
      - $ref: '#/components/parameters/PedidoId'
    post:
      tags: [Fluxos de Negócio]
      summary: Cancelar pedido
      description: |
        Disponível apenas para perfil `cliente` dono do pedido.
        Regra: somente pedido com status `EM_ANALISE` pode ser cancelado.
      operationId: cancelarPedido
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pedido cancelado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidoResponse'
        '401':
          description: Não autorizado
        '404':
          description: Pedido não encontrado
        '422':
          description: Regra de negócio violada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensagemErro'
  /pedidos/{id}/executar-contrato:
    parameters:
      - $ref: '#/components/parameters/PedidoId'
    post:
      tags: [Fluxos de Negócio]
      summary: Executar contrato por pedido
      description: |
        Disponível para perfis `empresa` e `banco`.
        Executa o contrato vinculado ao pedido e reflete no status do automóvel.
      operationId: executarContratoPorPedido
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Contrato executado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContratoResponse'
        '401':
          description: Não autorizado
        '404':
          description: Pedido/Contrato não encontrado
        '422':
          description: Regra de negócio violada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensagemErro'
...
  /contratos/{id}/executar:
    parameters:
      - $ref: '#/components/parameters/ContratoId'
    post:
      tags: [Contratos]
      summary: Executar contrato
      description: |
        Executa o contrato e atualiza o status do automóvel do pedido:
        - `CLIENTE` -> `ALUGADO`
        - `EMPRESA` ou `BANCO` -> `VINCULADO`
      operationId: executarContrato
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Contrato executado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContratoResponse'
        '401':
          description: Não autorizado
        '404':
          description: Contrato não encontrado
        '422':
          description: Regra de negócio violada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensagemErro'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token retornado em `POST /auth/login`
  parameters:
...
    PedidoId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    PedidoStatusPath:
      name: status
      in: path
      required: true
      schema:
        type: string
...
    ContratoId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    PedidoStatus:
      type: string
      enum: [EM_ANALISE, APROVADO, REJEITADO, CANCELADO]
    PedidoRequest:
      type: object
      required: [clienteId, automovelId, dataInicio, dataFim]
      properties:
        clienteId:
          type: integer
          format: int64
        automovelId:
          type: integer
          format: int64
        dataInicio:
          type: string
          format: date
        dataFim:
          type: string
          format: date
        status:
          type: string
          description: Campo opcional; na criação o backend força EM_ANALISE.
    PedidoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        numeroProtocolo:
          type: string
        clienteId:
          type: integer
          format: int64
        automovelId:
          type: integer
          format: int64
        dataInicio:
          type: string
          format: date
        dataFim:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/PedidoStatus'
...
    TipoPropriedade:
      type: string
      enum: [CLIENTE, EMPRESA, BANCO]
...
    ContratoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        pedidoId:
          type: integer
          format: int64
        automovelId:
          type: integer
          format: int64
        tipoPropriedade:
          $ref: '#/components/schemas/TipoPropriedade'
        dataAssinatura:
          type: string
          format: date
...
    MensagemErro:
      type: object
      properties:
        mensagem:
          type: string